Moved svg files to Assets folder
